ParticipantFinished endpoint
